Dropbox/doodrop/source/backend/routes/requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
pymongo==4.4.0
gunicorn==20.1.0